long_short_term_memory/requirements.txt
pytorch==1.12.1
lightning==2.0.1.post0
tensorboard== 2.12.2